chore(web) : web apis 경로 수정에 따른 각 파일들 수정 #205
