+++ apps/web/src/components/pages/order-page.tsx
@@ -10,7 +10,8 @@ import { useForm, type SubmitHandler } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { formSchema } from '@/validate/order-form-schema';
 import NonCustomerOrderForm from '@/components/order/non-customer-order-form';
-import { getOrderRound, insertOrders, searchBankList, searchDeliveryList } from '@/services/apis';
+import { insertOrders, searchBankList, searchDeliveryList } from '@/services/apis';
+import { getOrderRound } from '@/services/order-round-apis';
 import type { FormSchema } from '@/validate/order-form-schema';
 import useToast from '@/hooks/useToast';
 import { useParams } from '@tanstack/react-router';
