+++ apps/web/src/components/pages/products-page.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState } from 'react';
 import BreadSearch from '../projects/bread-search';
 import type { BreadProps, OrderRoundBreads } from '@/interface/bread-interface';
 import { useQuery } from '@tanstack/react-query';
-import { getOrderRoundNow, searchBreadList } from '@/services/apis';
+import { getOrderRoundNow } from '@/services/order-round-apis';
+import { searchBreadList } from '@/services/apis';
 import ProductsLoading from '../projects/products-loading';
 import BreadCardDetail from '../projects/bread-card-detail';
 import { AlertDialog } from '@appabbang/ui';
