+++ apps/web/src/services/common-apis.ts
@@ -1,0 +1,11 @@
+import axios from 'axios';
+
+/** axios 생성한 것을 컴포넌트에서 사용함. */
+export const client = axios.create({
+  baseURL: 'http://localhost:4000',
+  timeout: 10000,
+  headers: {
+    'Content-Type': 'application/json',
+    Accept: 'application/json',
+  },
+});
